+++ src/main/java/backend/siptis/service/report/activities/ActivitiesReportService.java
@@ -1,4 +1,4 @@
-package backend.siptis.service.report;
+package backend.siptis.service.report.activities;
 
 import backend.siptis.commons.ServiceAnswer;
