+++ src/main/java/backend/siptis/service/report/projects/ProjectReportService.java
@@ -1,4 +1,4 @@
-package backend.siptis.service.report;
+package backend.siptis.service.report.projects;
 
 import backend.siptis.commons.ServiceAnswer;
 
@@ -6,4 +6,6 @@ public interface ProjectReportService {
     ServiceAnswer getProjectTribunalReport();
 
     ServiceAnswer getProjectCareerReport();
+
+    ServiceAnswer getCompleteProjectReport();
 }
