+++ src/main/java/backend/siptis/service/report/projects/ProjectReportServiceImpl.java
@@ -1,4 +1,4 @@
-package backend.siptis.service.report;
+package backend.siptis.service.report.projects;
 
 import backend.siptis.commons.ServiceAnswer;
 import backend.siptis.commons.ServiceMessage;
@@ -7,8 +7,8 @@ import backend.siptis.model.entity.project_management.Project;
 import backend.siptis.model.entity.user_data.UserCareer;
 import backend.siptis.model.repository.project_management.ProjectRepository;
 import backend.siptis.service.cloud.CloudManagementService;
-import backend.siptis.service.report.geneartion_tools.ProjectByCareerReportTool;
-import backend.siptis.service.report.geneartion_tools.TribunalProjectReportTool;
+import backend.siptis.service.report.projects.generation_tools.ProjectByCareerReportTool;
+import backend.siptis.service.report.projects.generation_tools.TribunalProjectReportTool;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -50,6 +50,11 @@ public class ProjectReportServiceImpl implements ProjectReportService {
         String fileName = ProjectByCareerReportTool.generateReport(sistemas, informatica, combined);
         String key = cloud.uploadReportToCloud(fileName);
         return ServiceAnswer.builder().serviceMessage(ServiceMessage.DOCUMENT_GENERATED).data(key).build();
+    }
+
+    public ServiceAnswer getCompleteProjectReport() {
+        List<Project> projects = projectRepository.findAll();
+        return null;
     }
 
     private boolean isCareer(Project project, backend.siptis.commons.UserCareer compareCareer) {
