+++ src/main/java/backend/siptis/service/report/tribunals/general_tools/ReportServiceImpl.java
@@ -1,4 +1,4 @@
-package backend.siptis.service.report;
+package backend.siptis.service.report.tribunals.general_tools;
 
 import backend.siptis.auth.entity.SiptisUser;
 import backend.siptis.commons.ServiceAnswer;
@@ -6,7 +6,7 @@ import backend.siptis.commons.ServiceMessage;
 import backend.siptis.model.repository.auth.SiptisUserRepository;
 import backend.siptis.model.repository.project_management.ProjectRepository;
 import backend.siptis.service.cloud.CloudManagementService;
-import backend.siptis.service.report.geneartion_tools.GeneralTribunalReportTool;
+import backend.siptis.service.report.tribunals.GeneralTribunalReportTool;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
